[INJIMOB-3390] refactor: expose fromValue in public enums for easy integration
