+++ Sources/OpenID4VP/Constants/ContentEncryptionAlgorithm.swift
@@ -1,0 +1,7 @@
+public enum ContentEncryptionAlgorithm : String, Codable, CaseIterable {
+    case A256GCM
+    
+    public static func fromValue(_ value: String) -> ContentEncryptionAlgorithm? {
+        return ContentEncryptionAlgorithm(rawValue: value)
+    }
+}
