--- Sources/OpenID4VP/Constants/EncryptionConstants.swift
@@ -1,3 +1,0 @@
-public enum ContentEncryptionAlgorithm : String, Codable, CaseIterable {
-    case A256GCM
-}
