+++ Sources/OpenID4VP/Constants/KeyManagementAlgorithm.swift
@@ -1,0 +1,7 @@
+public enum KeyManagementAlgorithm : String, Codable, CaseIterable {
+    case ecdhEs = "ECDH-ES"
+    
+    public static func fromValue(_ value: String) -> KeyManagementAlgorithm? {
+        return KeyManagementAlgorithm(rawValue: value)
+    }
+}
