+++ Sources/OpenID4VP/Constants/ProofPurpose.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public enum ProofPurpose: String, Codable {
+    case vpProofPurpose = "authentication"
+}
