+++ Sources/OpenID4VP/Constants/RequestSigningAlgorithm.swift
@@ -1,0 +1,9 @@
+public enum RequestSigningAlgorithm : String, Codable, CaseIterable {
+    case edDsa = "EdDSA"
+    
+    public static func fromValue(_ value: String) -> RequestSigningAlgorithm? {
+        return RequestSigningAlgorithm(rawValue: value)
+    }
+}
+
+
