+++ Sources/OpenID4VP/Constants/ResponseMode.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum ResponseMode: String {
+    case directPost = "direct_post"
+    case directPostJwt = "direct_post.jwt"
+}
