+++ Sources/OpenID4VP/Constants/ResponseType.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum ResponseType: String {
+    case vp_token = "vp_token"
+}
