+++ Sources/OpenID4VP/Constants/VPFormatType.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum VPFormatType: String, Codable {
+    case ldp_vp = "ldp_vp"
+    case mso_mdoc = "mso_mdoc"
+}
